chore: moves src/common/container.ts and src/common/atlas/generatePassword.ts into src/helpers and src/helpers/packageInfo.ts into src/common
