+++ src/common/packageInfo.ts
