+++ src/helpers/generatePassword.ts
