+++ src/index.ts
@@ -5,7 +5,7 @@ import { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
 import { config } from "./common/config.js";
 import { Session } from "./common/session.js";
 import { Server } from "./server.js";
-import { packageInfo } from "./helpers/packageInfo.js";
+import { packageInfo } from "./common/packageInfo.js";
 import { Telemetry } from "./telemetry/telemetry.js";
 import { createEJsonTransport } from "./helpers/EJsonTransport.js";
